added modal url for projects create and export modals
